Optimise code further with better folder structure of files
