--- app/lib/features/spaces/model/space_permission_levels.dart
@@ -1,1 +1,0 @@
-enum PermissionLevel { admin, moderator, member }
